// server/src/accommodations/dto/update-accommodation.dto.ts
export class UpdateAccommodationDto {
  name?: string;
  city?: string;
  province?: string;
  type?: string;
  category?: string;
  description?: string;
  image?: string;
  images?: string;
  rating?: number;
  priceRange?: string;
  coordinates?: string;
  address?: string;
  phone?: string;
  email?: string;
  website?: string;
  amenities?: string;
  capacity?: number;
  breakfast?: boolean;
  checkInTime?: string;
  checkOutTime?: string;
  cancellationPolicy?: string;
  totalRooms?: number;
  availability?: string;
  bookingUrl?: string;
}